pyproject.toml
[project]
name = "domashka"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
requires-python = ">=3.14"
dependencies = [
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
lint = [
    "flake8 (>=7.3.0,<8.0.0)",
    "black (>=26.3.1,<27.0.0)",
    "isort (>=8.0.1,<9.0.0)",
    "mypy (>=1.19.1,<2.0.0)"
]
dev = [
    "pytest-cov (>=7.1.0,<8.0.0)"
]

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = "venv"
mypy_path = "src"
explicit_package_bases = true

[tool.black]
line-length = 119
exclude = '''
(
  /(
      \.eggs         # Исключить несколько общих каталогов
    | \.git          # в корне проекта
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py           # Также отдельно исключить файл с именем foo.py
                     # в корне проекта
)
'''

[tool.isort]
line_length = 119